Move the file within the amitkrishna folder and deleting from the previous place
