+++ 02_bubble/amitkrishna/bubble.go
@@ -13,19 +13,35 @@ func main() {
 }
 
 func bubble(s []int) []int {
-
-	end := len(s) - 1
+	cs1 := make([]int, len(s))
+	copy(cs1, s)
+	end := len(cs1) - 1
 	for {
 		if end == 0 {
 			break
 		}
-
-		for i := 0; i < len(s)-1; i++ {
-			if s[i] > s[i+1] {
-				s[i], s[i+1] = s[i+1], s[i]
+		for i := 0; i < len(cs1)-1; i++ {
+			if cs1[i] > cs1[i+1] {
+				cs1[i], cs1[i+1] = cs1[i+1], cs1[i]
 			}
 		}
 		end--
 	}
-	return s
+	return cs1
+}
+
+func insertionsort(items []int) []int {
+	cs2 := make([]int, len(items))
+	copy(cs2, items)
+	var n = len(cs2)
+	for i := 1; i < n; i++ {
+		j := i
+		for j > 0 {
+			if cs2[j-1] > cs2[j] {
+				cs2[j-1], cs2[j] = cs2[j], cs2[j-1]
+			}
+			j--
+		}
+	}
+	return cs2
 }
